fix(docs): update
